+++ src/main/java/com/zerobase/foodlier/global/member/mail/service/VerificationCodeServiceImpl.java
@@ -1,9 +1,9 @@
-package com.zerobase.foodlier.module.member.member.mail.service;
+package com.zerobase.foodlier.global.member.mail.service;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.stereotype.Service;
 
-import static com.zerobase.foodlier.module.member.member.mail.constants.GenerateCodeConstants.*;
+import static com.zerobase.foodlier.global.member.mail.constants.GenerateCodeConstants.*;
 
 @Service
 public class VerificationCodeServiceImpl implements VerificationCodeService{
